test(onboard): split dashboard port helpers
